add logging at tools

+++ voximplant/tools.py
@@ -1,9 +1,12 @@
 # coding: utf-8
+import logging
 import time
 from django.db.models import Q, F
 from django.utils.timezone import now
 from . import models
 from . import api_client
+
+logger = logging.getLogger('voximplant.tool')
 
 
 class SendCallListException(Exception):
@@ -20,13 +23,18 @@ def scenarios_download():
     incoming_ids = set([i['scenario_id'] for i in result])
     exists_ids = set(models.Scenario.objects.filter(vox_id__isnull=False).values_list('vox_id', flat=True))
     deleted_ids = exists_ids - incoming_ids
-    models.Scenario.objects.filter(vox_id__in=deleted_ids).delete()
+    if deleted_ids:
+        models.Scenario.objects.filter(vox_id__in=deleted_ids).delete()
+        logger.info('Delete scenarios', extra={'deleted_ids': deleted_ids})
 
     for item_data in result:
-        scenario, _ = models.Scenario.objects.get_or_create(vox_id=item_data['scenario_id'])
+        scenario, created = models.Scenario.objects.get_or_create(vox_id=item_data['scenario_id'])
         scenario.name = item_data['scenario_name']
         scenario.modified = item_data['modified'] + 'Z'
         scenario.save()
+
+        create_or_update = 'Created' if created else 'Updated'
+        logger.info('%s scenario' % create_or_update, extra={'scenario': item_data, 'id': scenario.id})
 
 
 def apps_download():
@@ -35,13 +43,18 @@ def apps_download():
     incoming_ids = set([i['application_id'] for i in result])
     exists_ids = set(models.Application.objects.filter(vox_id__isnull=False).values_list('vox_id', flat=True))
     deleted_ids = exists_ids - incoming_ids
-    models.Application.objects.filter(vox_id__in=deleted_ids).delete()
+    if deleted_ids:
+        models.Application.objects.filter(vox_id__in=deleted_ids).delete()
+        logger.info('Delete applications', extra={'deleted_ids': deleted_ids})
 
     for item_data in result:
-        app, _ = models.Application.objects.get_or_create(vox_id=item_data['application_id'])
+        app, created = models.Application.objects.get_or_create(vox_id=item_data['application_id'])
         app.name = item_data['application_name']
         app.modified = item_data['modified'] + 'Z'
         app.save()
+
+        create_or_update = 'Created' if created else 'Updated'
+        logger.info('%s application' % create_or_update, extra={'application': item_data, 'id': app.id})
 
 
 def apps_upload():
@@ -52,6 +65,7 @@ def apps_upload():
         if result.get('application_id'):
             update_params['vox_id'] = result['application_id']
         models.Application.objects.filter(id=app.id).update(**update_params)
+        logger.info('Upload application', extra={'result': result, 'id': app.id})
 
 
 def rules_upload():
@@ -66,11 +80,14 @@ def rules_upload():
         if result.get('rule_id'):
             update_params['vox_id'] = result['rule_id']
         models.Rule.objects.filter(id=rule.id).update(**update_params)
+        logger.info('Upload rule', extra={'result': result, 'id': rule.id})
 
         if is_new:
             rule = models.Rule.objects.get(id=rule.id)
             for scenario in rule.scenarios.filter(vox_id__isnull=False):
-                api_client.scenario_bind_rule(scenario.vox_id, rule.vox_id, True)
+                result = api_client.scenario_bind_rule(scenario.vox_id, rule.vox_id, True)
+                logger.info('Bind rule with scenario',
+                             extra={'result': result, 'scenario_id': scenario.id, 'rule_id': rule.id})
 
 
 def rules_download():
@@ -80,14 +97,19 @@ def rules_download():
         incoming_ids = set([i['rule_id'] for i in result])
         exists_ids = set(models.Rule.objects.filter(vox_id__isnull=False).values_list('vox_id', flat=True))
         deleted_ids = exists_ids - incoming_ids
-        models.Rule.objects.filter(application=app, vox_id__in=deleted_ids).delete()
+        if deleted_ids:
+            models.Rule.objects.filter(application=app, vox_id__in=deleted_ids).delete()
+            logger.info('Delete rules', extra={'deleted_ids': deleted_ids})
 
         for item_data in result:
-            rule, _ = models.Rule.objects.get_or_create(vox_id=item_data['rule_id'], defaults={'application': app})
+            rule, created = models.Rule.objects.get_or_create(vox_id=item_data['rule_id'], defaults={'application': app})
             rule.name = item_data['rule_name']
             rule.pattern = item_data['rule_pattern']
             rule.modified = item_data['modified'] + 'Z'
             rule.save()
+
+            create_or_update = 'Created' if created else 'Updated'
+            logger.info('%s rule' % create_or_update, extra={'rule': item_data, 'id': rule.id})
 
             incoming_scenario_ids = set([s['scenario_id'] for s in item_data['scenarios']])
             exists_scenario_ids = set(s.vox_id for s in rule.scenarios.all())
@@ -96,11 +118,13 @@ def rules_download():
             deleted_scenarios = models.Scenario.objects.filter(vox_id__in=deleted_scenario_ids)
             for deleted_scenario in deleted_scenarios:
                 rule.scenarios.remove(deleted_scenario)
+                logger.info('Unbind rule and scenario', extra={'rule_id': rule.id, 'scenario_id': deleted_scenario.id})
 
             added_scenario_ids = exists_scenario_ids - incoming_scenario_ids
             added_scenarios = models.Scenario.objects.filter(vox_id__in=added_scenario_ids)
             for added_scenario in added_scenarios:
                 rule.scenarios.add(added_scenario)
+                logger.info('Bind rule and scenario', extra={'rule_id': rule.id, 'scenario_id': added_scenario.id})
 
 
 def scenarios_upload():
@@ -116,13 +140,19 @@ def scenarios_upload():
                 update_params['vox_id'] = result['scenario_id']
             models.Scenario.objects.filter(id=scenario.id).update(**update_params)
             scenario = models.Scenario.objects.get(id=scenario.id)
+            logger.info('Upload scenario', extra={'result': result, 'id': scenario.id})
 
         remote_rules = api_client.scenario_get_rules(scenario.vox_id)
         local_rules = set([r.vox_id for r in scenario.rules.all() if r.vox_id])
         all_rules = local_rules | remote_rules
         for rule_vox_id in all_rules:
             bind = rule_vox_id in local_rules
-            api_client.scenario_bind_rule(scenario.vox_id, rule_vox_id, bind)
+            result = api_client.scenario_bind_rule(scenario.vox_id, rule_vox_id, bind)
+
+            bind_or_unbind = 'Bind' if bind else 'Unbind'
+            rule = models.Rule.objects.get(vox_id=rule_vox_id)
+            logger.info('%s rule and scenario' % bind_or_unbind,
+                         extra={'result': result, 'scenario_id': scenario.id, 'rule_id': rule.id})
 
 
 def call_list_send(call_list_id: int, force=False):
@@ -137,15 +167,17 @@ def call_list_send(call_list_id: int, force=False):
     call_list.vox_id = result['list_id']
     call_list.started = now()
     call_list.save()
+    logger.info('Send call list', extra={'result': result, 'id': call_list.id})
 
 
 def call_list_stop(call_list_id: int):
     call_list = models.CallList.objects.get(id=call_list_id)
     assert call_list.vox_id
 
-    api_client.call_list_stop(call_list)
+    result = api_client.call_list_stop(call_list)
     call_list.canceled = now()
     call_list.save()
+    logger.info('Stop call list', extra={'result': result, 'id': call_list.id})
 
 
 def call_list_recover(call_list_id: int):
@@ -155,12 +187,13 @@ def call_list_recover(call_list_id: int):
     if not call_list.canceled:
         raise SendCallListException('Call list with id "%s" is not stopped' % call_list.id)
 
-    api_client.call_list_recover(call_list)
+    result = api_client.call_list_recover(call_list)
     call_list.canceled = None
     call_list.save()
+    logger.info('Recover call list', extra={'result': result, 'id': call_list.id})
 
 
-def call_list_download(call_list_id: int, verbosity: int = 1):
+def call_list_download(call_list_id: int):
     call_list = models.CallList.objects.get(id=call_list_id)
     if not call_list.vox_id:
         raise DownloadCallListException('Call list with id "%s" have not vox_id' % call_list.id)
@@ -178,9 +211,10 @@ def call_list_download(call_list_id: int, verbosity: int = 1):
 
     call_list.downloaded = now()
     call_list.save()
+    logger.info('Got call list state', extra={'result': result, 'id': call_list.id})
 
 
-def call_lists_check(infinitely: bool = False, sleep_sec: int = 10, verbosity: int = 1):
+def call_lists_check(infinitely: bool = False, sleep_sec: int = 10, download_handler=None):
     while True:
         uncompleted_ids = set(
             models.CallListPhone.objects
@@ -188,22 +222,16 @@ def call_lists_check(infinitely: bool = False, sleep_sec: int = 10, verbosity: i
                 .filter(call_list__started__isnull=False, completed__isnull=True)
                 .filter(Q(call_list__canceled__isnull=True) | Q(call_list__canceled__gte=F('call_list__downloaded')))  # make last check callist state after canceling
                 .values_list('call_list__id', flat=True))
-        if verbosity > 1:
-            print('Started and uncompleted call lists: ', len(uncompleted_ids))
+        logger.debug('Checked uncompleted call lists', extra={'uncompleted_ids': uncompleted_ids})
 
         for call_list_id in uncompleted_ids:
-            if verbosity > 1:
-                print('Downloading call list with id "%s"... ' % call_list_id)
-
-            call_list_download(call_list_id, verbosity=verbosity)
-
-            if verbosity > 1:
-                print('complete.')
+            logger.debug('Downloading call list', extra={'id': call_list_id})
+            call_list_download(call_list_id)
+            if download_handler:
+                download_handler(call_list_id)
 
         if infinitely:
-            if verbosity > 2:
-                print('Sleep', sleep_sec)
-
+            logger.debug('Checking call list sleep...', extra={'sleep_sec': sleep_sec})
             time.sleep(sleep_sec)
         else:
             return
